# server/src/public/swagger.json
{
      "openapi": "3.0.0",
      "info": {
            "title": "MyGame-V1",
            "description": "a board game build base on NestJs and NextJs",
            "contact": {
                  "name": "pham vinh nhan",
                  "email": "heaty566@gmail.com"
            },
            "license": {
                  "name": "Apache 2.0",
                  "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
            },
            "version": "0.0.1"
      },
...
      "paths": {
...
            "/api/admin/users": {
                  "get": {
                        "tags": ["admin"],
                        "description": "get all users information except some sensitive information",
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "responses": {
                              "200": {
                                    "description": "return array of user information",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "403": {
                                    "$ref": "#/components/responses/403"
                              }
                        }
                  }
            },
            "/api/admin/user-admin/:id": {
                  "put": {
                        "tags": ["admin"],
                        "description": "toggle user role from USER to ADMIN and vice versa",
                        "parameters": [{ "$ref": "#/components/parameters/token" }, { "$ref": "#/components/parameters/userId" }],
                        "responses": {
                              "200": {
                                    "description": "update user role successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "403": {
                                    "$ref": "#/components/responses/403"
                              },
                              "404": {
                                    "$ref": "#/components/responses/404"
                              }
                        }
                  }
            },
            "/api/admin/user-status/:id": {
                  "put": {
                        "tags": ["admin"],
                        "description": "toggle user status from true to false and vice versa (block user)",
                        "parameters": [{ "$ref": "#/components/parameters/token" }, { "$ref": "#/components/parameters/userId" }],
                        "responses": {
                              "200": {
                                    "description": "update user status successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "403": {
                                    "$ref": "#/components/responses/403"
                              },
                              "404": {
                                    "$ref": "#/components/responses/404"
                              }
                        }
                  }
            }
      },
      "components": {
            "responses": {
                  "401": {
                        "description": "user has not a re-token",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
...
                  "403": {
                        "description": "user has to be an administrator",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
                  "404": {
                        "description": "entity does not exist in database",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
...
            },
            "schemas": {
                  "CommonApiResponse": {
                        "type": "object",
                        "properties": {
                              "data": {
                                    "type": "object",
                                    "description": "it type any base on your api"
                              },
                              "message": {
                                    "type": "string",
                                    "example": "this is an error message"
                              },
                              "details": {
                                    "type": "object",
                                    "properties": {
                                          "field": {
                                                "type": "string",
                                                "description": "field will be the name of invalid field",
                                                "example": "should not be empty"
                                          }
                                    }
                              }
                        }
                  },
...
            },
...
            "parameters": {
                  "token": {
                        "name": "token",
                        "in": "cookie",
                        "required": true,
                        "style": "form",
                        "schema": {
                              "type": "object",
                              "properties": {
                                    "re-token": {
                                          "type": "string",
                                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                    },
                                    "auth-token": {
                                          "type": "string",
                                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                    }
                              }
                        }
                  },
                  "userId": {
                        "name": "id",
                        "description": "user id",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "schema": {
                              "type": "string",
                              "example": "123456789aa"
                        }
                  },
...
            }
      }
}